[patch 27] Accepted patches from freebsd port. 	https://sourceforge.net/p/rezound/patches/27/ Also bumped c++0x to c++11

+++ rezound/src/backend/CPortAudioSoundPlayer.cpp
@@ -73,7 +73,7 @@ void CPortAudioSoundPlayer::initialize()
 		// open a PortAudio stream
 #ifdef ENABLE_PORTAUDIO_V19
 		PaStreamParameters output = { gPortAudioOutputDevice, 
-			gDesiredOutputChannelCount, 
+			static_cast<int>(gDesiredOutputChannelCount),
 			sampleFormat,
 			Pa_GetDeviceInfo(gPortAudioOutputDevice)->defaultLowOutputLatency ,
 			NULL};
